Rename and move abstract classes
