+++ pipelex/cogt/inference/inference_worker_abstract.py
@@ -8,7 +8,7 @@ from typing import Optional
 from pipelex.cogt.inference.inference_report_delegate import InferenceReportDelegate
 
 
-class InferenceReporterAbstract(ABC):
+class InferenceWorkerAbstract(ABC):
     def __init__(
         self,
         report_delegate: Optional[InferenceReportDelegate] = None,
