+++ pipelex/core/pipe_library.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Optional
 from pydantic import Field, RootModel
 from typing_extensions import override
 
-from pipelex.core.pipe import PipeAbstract
+from pipelex.core.pipe_abstract import PipeAbstract
 from pipelex.core.pipe_provider_abstract import PipeProviderAbstract
 from pipelex.exceptions import ConceptError, ConceptLibraryConceptNotFoundError, PipeLibraryError, PipeLibraryPipeNotFoundError
 from pipelex.hub import get_concept_provider
