+++ pipelex/pipe_controllers/pipe_controller.py
@@ -7,7 +7,7 @@ from typing import Optional
 
 from typing_extensions import override
 
-from pipelex.core.pipe import PipeAbstract
+from pipelex.core.pipe_abstract import PipeAbstract
 from pipelex.core.pipe_output import PipeOutput
 from pipelex.core.pipe_run_params import PipeRunParams
 from pipelex.core.working_memory import WorkingMemory
