+++ pipelex/pipe_works/pipe_job_factory.py
@@ -4,7 +4,7 @@
 
 from typing import Optional
 
-from pipelex.core.pipe import PipeAbstract
+from pipelex.core.pipe_abstract import PipeAbstract
 from pipelex.core.pipe_run_params import PipeRunParams
 from pipelex.core.working_memory import WorkingMemory
 from pipelex.mission.job_metadata import JobMetadata
